requirements.txt
Django==2.2.8
djangorestframework==3.13.1
djongo==1.3.6
pymongo==3.12.3